O4C_Python/BackEnd/requirements.txt
astroid==1.4.9
colorama==0.3.7
crypto==1.4.1
decorator==4.0.11
django==1.9
django-braces==1.11.0
django-cors-headers==2.0.1
django-oauth-toolkit==0.12.0
django-smtp-ssl==1.0
djangorestframework==3.4.6
djangorestframework-expiring-authtoken==0.1.4
ipython==5.2.1
ipython-genutils==0.1.0
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
mysqlclient==1.3.9
Naked==0.1.31
oauthlib==2.0.1
olefile==0.44
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.10
Pygments==2.2.0
pylint==1.6.5
pyreadline==2.1
PyYAML==3.12
requests==2.13.0
shellescape==3.4.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
virtualenv==15.1.0
wcwidth==0.1.7
wrapt==1.10.8
